Move all js code besides index.js into a folder in src
